pyproject.toml
[tool.poetry]
name = "price-manager"
version = "0.1.0"
description = ""
authors = ["VladimirKhrabrov <khrabrov.v@brandsgo.digital>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
psycopg2 = "^2.9.9"
xlrd = "^2.0.1"
openpyxl = "^3.1.5"
selenium = "^4.23.0"
webdriver-manager = "^4.0.1"
python-telegram-bot = "^21.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
